+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Use the timelines/tag endpoint to pull trending tag toots in addition to search endpoint
+* Log the Instance server info dict all at once
 
 ### v0.21.3
 * Add MastodonServer.fetchServerInfo() method to fetch all the available mastodon.v2.Instance data about the server instead of just MAU data